+++ src/components/contentRowContainer/contentRowContainer.component.ts
@@ -1,0 +1,30 @@
+import {Component, ChangeDetectionStrategy, OnInit} from '@angular/core';
+
+import {RenderableContentComponent} from '../renderableContent/renderableContent.component';
+
+/**
+ * Component that represents content row container
+ */
+@Component(
+{
+    selector: '[contentRowContainer]',
+    templateUrl: '../renderableContent/renderableContent.component.html',
+    standalone: true,
+    changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class ContentRowContainerSAComponent extends RenderableContentComponent implements OnInit
+{
+    //######################### public methods - implementation of OnInit #########################
+    
+    /**
+     * Initialize component
+     */
+    public ngOnInit(): void
+    {
+        this.contentRendererInnerStructure.contentRowContainer[this.contentRendererInnerStructure.contentRowContainer.length - 1].push(
+        {
+            renderableContent: this,
+            view: null,
+        });
+    }
+}