+++ src/components/footerRowContainer/footerRowContainer.component.ts
@@ -1,0 +1,30 @@
+import {Component, ChangeDetectionStrategy, OnInit} from '@angular/core';
+
+import {RenderableContentComponent} from '../renderableContent/renderableContent.component';
+
+/**
+ * Component that represents footer row container
+ */
+@Component(
+{
+    selector: '[footerRowContainer]',
+    templateUrl: '../renderableContent/renderableContent.component.html',
+    standalone: true,
+    changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class FooterRowContainerSAComponent extends RenderableContentComponent implements OnInit
+{
+    //######################### public methods - implementation of OnInit #########################
+    
+    /**
+     * Initialize component
+     */
+    public ngOnInit(): void
+    {
+        this.contentRendererInnerStructure.footerRowContainer.push(
+        {
+            renderableContent: this,
+            view: null,
+        });
+    }
+}